Extract form component
build bundle
